refactor code for createAuxImage to remove duplicate code blocks
