+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cli/menu/RebaseImage.java
@@ -4,30 +4,17 @@
 package com.oracle.weblogic.imagetool.cli.menu;
 
 import java.io.File;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.List;
 import java.util.Properties;
-import java.util.UUID;
 import java.util.concurrent.Callable;
 
-import com.oracle.weblogic.imagetool.api.model.CachedFile;
 import com.oracle.weblogic.imagetool.api.model.CommandResponse;
-import com.oracle.weblogic.imagetool.builder.BuildCommand;
-import com.oracle.weblogic.imagetool.installer.FmwInstallerType;
-import com.oracle.weblogic.imagetool.installer.InstallerType;
-import com.oracle.weblogic.imagetool.installer.MiddlewareInstall;
 import com.oracle.weblogic.imagetool.logging.LoggingFacade;
 import com.oracle.weblogic.imagetool.logging.LoggingFactory;
-import com.oracle.weblogic.imagetool.util.Constants;
-import com.oracle.weblogic.imagetool.util.DockerfileOptions;
 import com.oracle.weblogic.imagetool.util.Utils;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
-
-import static com.oracle.weblogic.imagetool.cachestore.CacheStoreFactory.cache;
 
 @Command(
     name = "rebase",
@@ -35,7 +22,7 @@ import static com.oracle.weblogic.imagetool.cachestore.CacheStoreFactory.cache;
     requiredOptionMarker = '*',
     abbreviateSynopsis = true
 )
-public class RebaseImage extends CommonOptions implements Callable<CommandResponse> {
+public class RebaseImage extends CommonCreateOptions implements Callable<CommandResponse> {
 
     private static final LoggingFacade logger = LoggingFactory.getLogger(RebaseImage.class);
 
@@ -44,9 +31,6 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
         logger.entering();
         Instant startTime = Instant.now();
 
-        String buildId = UUID.randomUUID().toString();
-        String tmpDir = null;
-        dockerfileOptions = new DockerfileOptions(buildId);
         String oldOracleHome;
         String oldJavaHome;
         String newOracleHome = null;
@@ -56,10 +40,7 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
         boolean modelOnly;
 
         try {
-
-            init(buildId);
-
-            tmpDir = getTempDirectory();
+            initializeOptions();
 
             if (sourceImage != null && !sourceImage.isEmpty()) {
                 logger.finer("IMG-0002", sourceImage);
@@ -67,7 +48,7 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
 
                 logger.info("IMG-0091", sourceImage);
                 Properties baseImageProperties = Utils.getBaseImageProperties(buildEngine, sourceImage,
-                    "/probe-env/inspect-image.sh", tmpDir);
+                    "/probe-env/inspect-image.sh", buildDir());
 
                 oldOracleHome = baseImageProperties.getProperty("oracleHome", null);
                 oldJavaHome = baseImageProperties.getProperty("javaHome", null);
@@ -84,7 +65,7 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
                 dockerfileOptions.setRebaseToTarget(true);
 
                 Properties baseImageProperties = Utils.getBaseImageProperties(buildEngine, targetImage,
-                    "/probe-env/inspect-image.sh", tmpDir);
+                    "/probe-env/inspect-image.sh", buildDir());
 
                 newOracleHome = baseImageProperties.getProperty("oracleHome", null);
                 newJavaHome = baseImageProperties.getProperty("javaHome", null);
@@ -117,55 +98,22 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
                 .setWdtModelHome(modelHome)
                 .setWdtModelOnly(modelOnly);
 
-            BuildCommand cmdBuilder = getInitialBuildCmd(tmpDir);
-
             if (dockerfileOptions.isRebaseToNew()) {
-
-                copyOptionsFromImage(fromImage, tmpDir);
-
-                CachedFile jdk = new CachedFile(InstallerType.JDK, jdkVersion);
-                Path installerPath = jdk.copyFile(cache(), tmpDir);
-                dockerfileOptions.setJavaInstaller(installerPath.getFileName().toString());
-
-                MiddlewareInstall install =
-                    new MiddlewareInstall(installerType, installerVersion, installerResponseFiles);
-                install.copyFiles(cache(), tmpDir);
-                dockerfileOptions.setMiddlewareInstall(install);
-
-                // resolve required patches
-                handlePatchFiles(installerType);
-
-                // If patching, patch OPatch first
-                if (applyingPatches() && shouldUpdateOpatch()) {
-                    installOpatchInstaller(tmpDir, opatchBugNumber);
-                }
-
-                Utils.setOracleHome(installerResponseFiles, dockerfileOptions);
-
-                // Set the inventory oraInst.loc file location (null == default location)
-                dockerfileOptions.setInvLoc(inventoryPointerInstallLoc);
-
-                // Set the inventory pointer
-                if (inventoryPointerFile != null) {
-                    Utils.setInventoryLocation(inventoryPointerFile, dockerfileOptions);
-                    Utils.copyLocalFile(Paths.get(inventoryPointerFile), Paths.get(tmpDir,"/oraInst.loc"));
-                } else {
-                    Utils.copyResourceAsFile("/response-files/oraInst.loc", tmpDir);
-                }
+                prepareNewImage();
             }
 
             // Create Dockerfile
-            String dockerfile = Utils.writeDockerfile(tmpDir + File.separator + "Dockerfile",
+            String dockerfile = Utils.writeDockerfile(buildDir() + File.separator + "Dockerfile",
                 "Rebase_Image.mustache", dockerfileOptions, dryRun);
 
             // add directory to pass the context
-            runDockerCommand(dockerfile, cmdBuilder);
+            runDockerCommand(dockerfile, getInitialBuildCmd(buildDir()));
         } catch (Exception ex) {
             return CommandResponse.error(ex.getMessage());
         } finally {
             if (!skipcleanup) {
-                Utils.deleteFilesRecursively(tmpDir);
-                Utils.removeIntermediateDockerImages(buildEngine, buildId);
+                Utils.deleteFilesRecursively(buildDir());
+                Utils.removeIntermediateDockerImages(buildEngine, buildId());
             }
         }
         Instant endTime = Instant.now();
@@ -174,37 +122,9 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
             return CommandResponse.success("IMG-0054");
         } else {
             return CommandResponse.success("IMG-0053",
-                Duration.between(startTime, endTime).getSeconds(), imageTag);
+                Duration.between(startTime, endTime).getSeconds(), imageTag());
         }
     }
-
-    @Override
-    String getInstallerVersion() {
-        return installerVersion;
-    }
-
-    @Option(
-        names = {"--type"},
-        description = "Installer type. Default: WLS. Supported values: ${COMPLETION-CANDIDATES}",
-        defaultValue = "wls"
-    )
-    private FmwInstallerType installerType;
-
-    @Option(
-        names = {"--version"},
-        description = "Installer version. Default: ${DEFAULT-VALUE}",
-        required = true,
-        defaultValue = Constants.DEFAULT_WLS_VERSION
-    )
-    private String installerVersion;
-
-    @Option(
-        names = {"--jdkVersion"},
-        description = "Version of server jdk to install. Default: ${DEFAULT-VALUE}",
-        required = true,
-        defaultValue = Constants.DEFAULT_JDK_VERSION
-    )
-    private String jdkVersion;
 
     @Option(
         names = {"--sourceImage"},
@@ -219,28 +139,4 @@ public class RebaseImage extends CommonOptions implements Callable<CommandRespon
     )
     private String targetImage;
 
-    @Option(
-        names = {"--installerResponseFile"},
-        split = ",",
-        description = "path to a response file. Override the default responses for the Oracle installer"
-    )
-    private List<Path> installerResponseFiles;
-
-    @Option(
-        names = {"--fromImage"},
-        description = "Docker image to use as base image."
-    )
-    private String fromImage;
-
-    @Option(
-        names = {"--inventoryPointerFile"},
-        description = "path to a user provided inventory pointer file as input"
-    )
-    private String inventoryPointerFile;
-
-    @Option(
-        names = {"--inventoryPointerInstallLoc"},
-        description = "path to where the inventory pointer file (oraInst.loc) should be stored in the image"
-    )
-    private String inventoryPointerInstallLoc;
 }
